+++ Server/Git/IGitToolsCommand.cs
@@ -1,0 +1,6 @@
+﻿namespace PrincipleStudios.ScaledGitApp.Git;
+
+public interface IGitToolsCommand<T> where T : Task
+{
+	T RunCommand(IGitToolsPowerShell pwsh);
+}
