+++ Server/Git/IGitToolsInvoker.cs
@@ -1,0 +1,7 @@
+﻿namespace PrincipleStudios.ScaledGitApp.Git;
+
+public interface IGitToolsInvoker
+{
+	Task RunCommand(IGitToolsCommand<Task> command);
+	Task<T> RunCommand<T>(IGitToolsCommand<Task<T>> command);
+}
